requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
html-to-markdown==1.3.2
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
nodeenv==1.9.1
outcome==1.3.0.post0
prisma==0.15.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
selenium==4.32.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
tomlkit==0.13.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
